5714 Split out QueryExecutorSplitQueriesTest
