+++ java/query/query-core/src/test/java/sleeper/query/core/rowretrieval/QueryExecutorTest.java
@@ -20,44 +20,24 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import sleeper.core.partition.PartitionTree;
 import sleeper.core.partition.PartitionsBuilder;
-import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.core.properties.table.TableProperties;
-import sleeper.core.range.Range.RangeFactory;
-import sleeper.core.range.Region;
 import sleeper.core.row.Row;
-import sleeper.core.row.testutils.InMemoryRowStore;
 import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.ByteArrayType;
 import sleeper.core.schema.type.IntType;
 import sleeper.core.schema.type.LongType;
 import sleeper.core.schema.type.StringType;
-import sleeper.core.statestore.FileReference;
-import sleeper.core.statestore.FileReferenceFactory;
-import sleeper.core.statestore.StateStore;
-import sleeper.core.statestore.testutils.InMemoryTransactionLogStateStore;
-import sleeper.core.statestore.testutils.InMemoryTransactionLogs;
-import sleeper.core.util.ObjectFactory;
 import sleeper.example.iterator.AdditionIterator;
 import sleeper.example.iterator.SecurityFilteringIterator;
-import sleeper.query.core.model.LeafPartitionQuery;
 import sleeper.query.core.model.Query;
 import sleeper.query.core.model.QueryException;
-import sleeper.query.core.model.QueryProcessingConfig;
 
-import java.io.IOException;
 import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Spliterators;
-import java.util.UUID;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
-import static java.util.Spliterator.IMMUTABLE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static sleeper.core.properties.table.TableProperty.AGGREGATION_CONFIG;
@@ -65,19 +45,9 @@ import static sleeper.core.properties.table.TableProperty.FILTERING_CONFIG;
 import static sleeper.core.properties.table.TableProperty.ITERATOR_CLASS_NAME;
 import static sleeper.core.properties.table.TableProperty.ITERATOR_CONFIG;
 import static sleeper.core.properties.table.TableProperty.QUERY_PROCESSOR_CACHE_TIMEOUT;
-import static sleeper.core.properties.table.TableProperty.TABLE_ID;
-import static sleeper.core.properties.table.TableProperty.TABLE_NAME;
-import static sleeper.core.properties.testutils.InstancePropertiesTestHelper.createTestInstanceProperties;
-import static sleeper.core.properties.testutils.TablePropertiesTestHelper.createTestTableProperties;
-import static sleeper.core.schema.SchemaTestHelper.createSchemaWithKey;
 import static sleeper.core.statestore.testutils.StateStoreUpdatesWrapper.update;
 
-public class QueryExecutorTest {
-    private final InstanceProperties instanceProperties = createTestInstanceProperties();
-    private final InMemoryRowStore rowStore = new InMemoryRowStore();
-    private final Schema schema = createSchemaWithKey("key", new LongType());
-    private final TableProperties tableProperties = createTestTableProperties(instanceProperties, schema);
-    private final StateStore stateStore = InMemoryTransactionLogStateStore.createAndInitialise(tableProperties, new InMemoryTransactionLogs());
+public class QueryExecutorTest extends QueryExecutorTestBase {
 
     @Nested
     @DisplayName("Query rows")
@@ -122,7 +92,7 @@ public class QueryExecutorTest {
         @Test
         void shouldNotFindRowOutsidePartitionRangeWhenFileContainsAnInactiveRow() throws Exception {
             // Given
-            update(stateStore).initialise(new PartitionsBuilder(schema)
+            update(stateStore).initialise(new PartitionsBuilder(tableProperties)
                     .rootFirst("root")
                     .splitToNewChildren("root", "L", "R", 5L)
                     .buildList());
@@ -495,209 +465,5 @@ public class QueryExecutorTest {
             // Then the rows that were added are not found
             assertThat(getRows(queryExecutor, queryAllRows())).isEmpty();
         }
-    }
-
-    @Nested
-    @DisplayName("Split queries into leaf partition queries")
-    class SplitQueries {
-
-        @Test
-        void shouldCreateNoSubqueriesWhenNoFilesArePresent() throws Exception {
-            // Given
-            Query query = queryRange(1L, 10L);
-
-            // When / Then
-            assertThat(executor().splitIntoLeafPartitionQueries(query))
-                    .isEmpty();
-        }
-
-        @Test
-        void shouldCreateOneSubqueryWithOneFileInOnePartition() throws Exception {
-            // Given
-            addRootFile("test.parquet", List.of(new Row(Map.of("key", 1L))));
-            Region region = range(0L, 10L);
-            Query query = queryRegions(region);
-
-            // When / Then
-            assertThat(executor().splitIntoLeafPartitionQueries(query))
-                    .usingRecursiveFieldByFieldElementComparatorIgnoringFields("subQueryId")
-                    .containsExactly(LeafPartitionQuery.builder()
-                            .parentQuery(query)
-                            .tableId(tableProperties.get(TABLE_ID))
-                            .subQueryId("ignored")
-                            .regions(List.of(region))
-                            .leafPartitionId("root")
-                            .partitionRegion(rootPartitionRegion())
-                            .files(List.of("test.parquet"))
-                            .build());
-        }
-
-        @Test
-        void shouldCreateOneSubqueryWithMultipleFilesInOnePartition() throws Exception {
-            // Given
-            addRootFile("file1.parquet", List.of(new Row(Map.of("key", 1L))));
-            addRootFile("file2.parquet", List.of(new Row(Map.of("key", 2L))));
-            addRootFile("file3.parquet", List.of(new Row(Map.of("key", 3L))));
-            Region region = range(0L, 5L);
-            Query query = queryRegions(region);
-
-            // When / Then
-            assertThat(executor().splitIntoLeafPartitionQueries(query))
-                    .usingRecursiveFieldByFieldElementComparatorIgnoringFields("subQueryId")
-                    .containsExactly(LeafPartitionQuery.builder()
-                            .parentQuery(query)
-                            .tableId(tableProperties.get(TABLE_ID))
-                            .subQueryId("ignored")
-                            .regions(List.of(region))
-                            .leafPartitionId("root")
-                            .partitionRegion(rootPartitionRegion())
-                            .files(List.of("file1.parquet", "file2.parquet", "file3.parquet"))
-                            .build());
-        }
-
-        @Test
-        void shouldCreateTwoSubqueriesWithTwoLeafPartitions() throws Exception {
-            // Given
-            PartitionTree tree = new PartitionsBuilder(tableProperties)
-                    .rootFirst("root")
-                    .splitToNewChildren("root", "left", "right", 5L)
-                    .buildTree();
-            update(stateStore).initialise(tree);
-            addPartitionFile("left", "left1.parquet", List.of(new Row(Map.of("key", 1L))));
-            addPartitionFile("left", "left2.parquet", List.of(new Row(Map.of("key", 1L))));
-            addPartitionFile("right", "right1.parquet", List.of(new Row(Map.of("key", 2L))));
-            addPartitionFile("right", "right2.parquet", List.of(new Row(Map.of("key", 2L))));
-            Region region = range(0L, 10L);
-            Query query = queryRegions(region);
-
-            // When / Then
-            assertThat(executor().splitIntoLeafPartitionQueries(query))
-                    .usingRecursiveFieldByFieldElementComparatorIgnoringFields("subQueryId")
-                    .containsExactlyInAnyOrder(
-                            LeafPartitionQuery.builder()
-                                    .parentQuery(query)
-                                    .tableId(tableProperties.get(TABLE_ID))
-                                    .subQueryId("ignored")
-                                    .regions(List.of(region))
-                                    .leafPartitionId("left")
-                                    .partitionRegion(tree.getPartition("left").getRegion())
-                                    .files(List.of("left1.parquet", "left2.parquet"))
-                                    .build(),
-                            LeafPartitionQuery.builder()
-                                    .parentQuery(query)
-                                    .tableId(tableProperties.get(TABLE_ID))
-                                    .subQueryId("ignored")
-                                    .regions(List.of(region))
-                                    .leafPartitionId("right")
-                                    .partitionRegion(tree.getPartition("right").getRegion())
-                                    .files(List.of("right1.parquet", "right2.parquet"))
-                                    .build());
-        }
-
-    }
-
-    private void addRootFile(String filename, List<Row> rows) {
-        addFile(fileReferenceFactory().rootFile(filename, rows.size()), rows);
-    }
-
-    private void addPartitionFile(String partitionId, String filename, List<Row> rows) {
-        addFile(fileReferenceFactory().partitionFile(partitionId, filename, rows.size()), rows);
-    }
-
-    private void addFile(FileReference fileReference, List<Row> rows) {
-        addFileMetadata(fileReference);
-        rowStore.addFile(fileReference.getFilename(), rows);
-    }
-
-    private void addFileMetadata(FileReference fileReference) {
-        update(stateStore).addFile(fileReference);
-    }
-
-    private QueryExecutor executor() throws Exception {
-        return executorAtTime(Instant.now());
-    }
-
-    private QueryExecutor executorAtTime(Instant time) throws Exception {
-        QueryExecutor executor = uninitialisedExecutorAtTime(time);
-        executor.init(time);
-        return executor;
-    }
-
-    private QueryExecutor uninitialisedExecutorAtTime(Instant time) {
-        return new QueryExecutor(ObjectFactory.noUserJars(), stateStore, tableProperties,
-                new InMemoryLeafPartitionRowRetriever(rowStore), time);
-    }
-
-    private List<Row> getRows(Query query) throws Exception {
-        return getRows(executor(), query);
-    }
-
-    private List<Row> getRows(QueryExecutor executor, Query query) {
-        try (var it = executor.execute(query)) {
-            return StreamSupport.stream(Spliterators.spliteratorUnknownSize(it, IMMUTABLE), false)
-                    .collect(Collectors.toUnmodifiableList());
-        } catch (QueryException | IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private Query.Builder query() {
-        return Query.builder().queryId(UUID.randomUUID().toString())
-                .tableName(tableProperties.get(TABLE_NAME));
-    }
-
-    private Query queryAllRows() {
-        return queryAllRowsBuilder().build();
-    }
-
-    private Query.Builder queryAllRowsBuilder() {
-        return query()
-                .regions(List.of(partitionTree().getRootPartition().getRegion()));
-    }
-
-    private Query queryRange(Object min, Object max) {
-        return queryRegions(range(min, max));
-    }
-
-    private Query queryRegions(Region... regions) {
-        return query()
-                .regions(List.of(regions))
-                .build();
-    }
-
-    private Region range(Object min, Object max) {
-        RangeFactory factory = new RangeFactory(tableProperties.getSchema());
-        return new Region(factory.createRange("key", min, max));
-    }
-
-    private PartitionTree partitionTree() {
-        return new PartitionTree(stateStore.getAllPartitions());
-    }
-
-    private FileReferenceFactory fileReferenceFactory() {
-        return FileReferenceFactory.from(stateStore);
-    }
-
-    private Region rootPartitionRegion() {
-        return new PartitionTree(stateStore.getAllPartitions()).getRootPartition().getRegion();
-    }
-
-    private static QueryProcessingConfig requestValueFields(String... fields) {
-        return QueryProcessingConfig.builder()
-                .requestedValueFields(List.of(fields))
-                .build();
-    }
-
-    private static QueryProcessingConfig applyIterator(Class<?> iteratorClass) {
-        return QueryProcessingConfig.builder()
-                .queryTimeIteratorClassName(iteratorClass.getName())
-                .build();
-    }
-
-    private static QueryProcessingConfig applyIterator(Class<?> iteratorClass, String config) {
-        return QueryProcessingConfig.builder()
-                .queryTimeIteratorClassName(iteratorClass.getName())
-                .queryTimeIteratorConfig(config)
-                .build();
     }
 }
